refactor: Reorganizar estrutura de rotas com route group
